AR: Adding support to offload wake-up rule configuration to FM ucode.
Auto Response module supports the configuration of wake-up rules
so that system can be waken-up when the specofic scenario occurs.

Patch provides support to offload wake-up rules configuration to FM code.
Mainlty, It contains two interfaces. One to interact with user space
application to recieve wake-up rules configuration and another to interact
with FM ucode to offload the same configuration when triggered for
deep sleep event.

Change-Id: I7202144b6221b2647620a3e96a95828527e85bbe
Signed-off-by: Sunil Kumar Kori <Sunil.Kori@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/5418
Reviewed-by: Nipun Gupta <Nipun.Gupta@freescale.com>
Reviewed-by: Rajan Gupta <rajan.gupta@freescale.com>
Tested-by: Rajan Gupta <rajan.gupta@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/9244